Update the reference to data.govt.nz
This in the long term will allow us to create a new feature on data.govt.nz that can surface a reuse example when someone visits the dataset on data.govt.nz.

Also, we might need to set a controlled set of values for the licence eg. do we use `Creative Commons Attribution 4.0 International licence`, or `CC BY 4.0` or a URI to denote a licence?